:sparkles: chore(api): move get user-drop-item-poi repository code from ItemRepository to ItemLocationRepository
