+++ backend/streetdrop-api/src/main/java/com/depromeet/domains/item/repository/QueryDslItemRepository.java
@@ -1,11 +1,9 @@
 package com.depromeet.domains.item.repository;
 
 import com.depromeet.domains.item.dao.ItemDao;
-import com.depromeet.domains.user.dao.UserItemPointDao;
 
 import java.util.List;
 
 public interface QueryDslItemRepository {
     List<ItemDao> findByUserId(Long userId, long lastCursor);
-    List<UserItemPointDao> findByUserId(Long userId);
 }
